+++ Test/MSBuild.Tests.ps1
@@ -68,7 +68,7 @@ BeforeAll {
             $Project
         )
 
-        $script:path = Join-Path -Path $script:testRoot -ChildPath 'BuildRoot\project.msbuild'
+        $script:path = Join-Path -Path (Get-BuildRoot) -ChildPath 'project.msbuild'
         New-Item -Path $script:path -ItemType 'File' -Force
         $Project | Set-Content -Path $script:path -Force
         $script:path = $script:path | Split-Path -Leaf
@@ -254,6 +254,11 @@ BeforeAll {
         Join-Path -Path (Get-BuildRoot) -ChildPath '*.build' | Should -Exist
     }
 
+    function ThenNoErrors
+    {
+        $Global:Error | Should -BeNullOrEmpty
+    }
+
     function ThenNuGetPackagesRestored
     {
         foreach( $item in $script:path )
@@ -382,9 +387,13 @@ BeforeAll {
 
     function ThenSpecificNuGetVersionInstalled
     {
-        $nuGetPackageVersion = 'NuGet.CommandLine.{0}' -f $script:nuGetVersion
+        $nuGetPackageVersion = "NuGet.CommandLine.$($script:nuGetVersion)"
 
-        Join-Path -Path (Get-BuildRoot) -ChildPath ('packages\{0}' -f $nugetPackageVersion) | Should -Exist
+        Get-ChildItem -Path (Join-Path -Path (Get-BuildRoot) -ChildPath 'packages') |
+            Out-String |
+            Write-Debug
+
+        Join-Path -Path (Get-BuildRoot) -ChildPath "packages\$($nugetPackageVersion)" | Should -Exist
     }
 
     function ThenTaskFailed
@@ -422,6 +431,7 @@ Describe 'MSBuild' {
     It 'should compile project' {
         GivenAProjectThatCompiles
         WhenRunningTask -AsDeveloper
+        ThenNoErrors
         ThenNuGetPackagesRestored
         ThenProjectsCompiled
         ThenAssembliesAreNotVersioned
@@ -431,6 +441,7 @@ Describe 'MSBuild' {
     It 'should build multiple projects' {
         GivenProjectsThatCompile
         WhenRunningTask -AsDeveloper
+        ThenNoErrors
         ThenNuGetPackagesRestored
         ThenProjectsCompiled
         ThenAssembliesAreNotVersioned
@@ -439,6 +450,7 @@ Describe 'MSBuild' {
     It 'should version with build metadata' {
         GivenAProjectThatCompiles
         WhenRunningTask -AsBuildServer -AtVersion '1.5.9-rc.45+1034.master.deadbee'
+        ThenNoErrors
         ThenNuGetPackagesRestored
         ThenProjectsCompiled
         ThenAssembliesAreVersioned -ProductVersion '1.5.9-rc.45+1034.master.deadbee' -FileVersion '1.5.9'
@@ -474,62 +486,73 @@ Describe 'MSBuild' {
     It 'should run clean target in clean mode' {
         GivenAProjectThatCompiles
         WhenRunningTask -AsDeveloper
+        ThenNoErrors
         ThenProjectsCompiled
         WhenRunningTask -InCleanMode -AsDeveloper
+        ThenNoErrors
         ThenBinsAreEmpty
     }
 
     It 'should customize output verbosity' {
         GivenAProjectThatCompiles
         WhenRunningTask -AsDeveloper -WithParameter @{ 'Verbosity' = 'q'; }
+        ThenNoErrors
         ThenOutputIsEmpty
     }
 
     It 'should use minimal verbosity' {
         GivenAProjectThatCompiles
         WhenRunningTask -AsDeveloper
+        ThenNoErrors
         ThenOutputIsMinimal
     }
 
     It 'should pass extra build properties' {
         GivenAProjectThatCompiles
         WhenRunningTask -AsDeveloper -WithParameter @{ 'Property' = @( 'Fubar=Snafu' ) ; 'Verbosity' = 'diag' }
+        ThenNoErrors
         ThenOutput -Contains 'Fubar=Snafu'
     }
 
     It 'should pass custom arguments' {
         GivenAProjectThatCompiles
         WhenRunningTask -AsDeveloper -WithParameter @{ 'Argument' = @( '/nologo', '/version' ) }
+        ThenNoErrors
         ThenOutput -Contains '\d+\.\d+\.\d+\.\d+'
     }
 
     It 'should pass a single custom argument' {
         GivenAProjectThatCompiles
         WhenRunningTask -AsDeveloper -WithParameter @{ 'Argument' = @( '/version' ) }
+        ThenNoErrors
         ThenOutput -Contains '\d+\.\d+\.\d+\.\d+'
     }
 
     It 'should multi-CPU build' {
         GivenAProjectThatCompiles
         WhenRunningTask -AsDeveloper -WithParameter @{ 'Verbosity' = 'n' }
+        ThenNoErrors
         ThenOutput -Contains '\n\ {5}\d>'
     }
 
     It 'should pass CPU argument' {
         GivenAProjectThatCompiles
         WhenRunningTask -AsDeveloper -WithParameter @{ 'CpuCount' = 1; 'Verbosity' = 'n' }
+        ThenNoErrors
         ThenOutput -DoesNotContain '^\ {5}\d>'
     }
 
     It 'should use custom output directory' {
         GivenAProjectThatCompiles
         WhenRunningTask -AsDeveloper -WithParameter @{ 'OutputDirectory' = '.myoutput' }
+        ThenNoErrors
         ThenProjectsCompiled -To '.myoutput'
     }
 
     It 'should build custom targets' {
         GivenCustomMSBuildScriptWithMultipleTargets
         WhenRunningTask -AsDeveloper -WithParameter @{ 'Target' = 'clean','build' ; 'Verbosity' = 'diag' }
+        ThenNoErrors
         ThenBothTargetsRun
     }
 
@@ -542,6 +565,7 @@ Describe 'MSBuild' {
     }
 
     It 'should use custom version of MSBuild' {
+        ThenNoErrors
         GivenProject @"
 <?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
@@ -597,6 +621,7 @@ Describe 'MSBuild' {
         }
         GivenVersion $script:version
         WhenRunningTask -AsDeveloper -WithParameter @{ 'NoMaxCpuCountArgument' = $true ; 'NoFileLogger' = $true; }
+        ThenNoErrors
         ThenOutput -Contains ([regex]::Escape($msbuildRoot.TrimEnd('\')))
     }
 
@@ -609,6 +634,7 @@ Describe 'MSBuild' {
 </Project>
 "@
         WhenRunningTask -AsDeveloper -WithParameter @{ 'NoMaxCpuCountArgument' = $true; Verbosity = 'diag' }
+        ThenNoErrors
         ThenOutput -Contains ('MSBuildNodeCount = 1')
     }
 
@@ -621,6 +647,7 @@ Describe 'MSBuild' {
 </Project>
 "@
         WhenRunningTask -AsDeveloper -WithParameter @{ 'NoFileLogger' = $true }
+        ThenNoErrors
         ThenOutputNotLogged
     }
 
@@ -628,6 +655,7 @@ Describe 'MSBuild' {
         GivenProjectsThatCompile
         GivenNuGetVersion '5.9.3'
         WhenRunningTask -AsDeveloper
+        ThenNoErrors
         ThenSpecificNuGetVersionInstalled
         ThenNuGetPackagesRestored
     }
@@ -635,6 +663,7 @@ Describe 'MSBuild' {
     It 'should use 64-bit MSBuild' {
         GivenProject $procArchProject
         WhenRunningTask -AsDeveloper
+        ThenNoErrors
         ThenOutput -Contains 'PROCESSOR_ARCHITECTURE = AMD64'
     }
 
@@ -642,6 +671,7 @@ Describe 'MSBuild' {
         GivenProject $procArchProject
         GivenUse32BitIs 'true'
         WhenRunningTask -AsDeveloper
+        ThenNoErrors
         ThenOutput -Contains 'PROCESSOR_ARCHITECTURE = x86'
     }
 }